requirements.txt
-e .
pytest
boto3
boto3-stubs==1.16.43.0
aws-cdk.aws-lambda
aws-cdk.aws_apigateway
aws-cdk.aws_dynamodb
aws-cdk.aws_ec2
cdk-dynamo-table-viewer
